Separate sidekiq integration into separate files
