+++ lib/raven/integrations/sidekiq/sidekiq_cleanup_middleware.rb
@@ -1,0 +1,11 @@
+module Raven
+  class SidekiqCleanupMiddleware
+    def call(_worker, job, queue)
+      Raven.context.transaction.push "Sidekiq/#{job['class']}"
+      Raven.extra_context(:sidekiq => job.merge("queue" => queue))
+      yield
+      Context.clear!
+      BreadcrumbBuffer.clear!
+    end
+  end
+end
